Adding Styles To gh-pages
Adding styles and pics and stuff to gh-pages. I have overridden
_layouts/default.html to show an OpenGraph image tag. Not sure if
this is nessasarry?... It looks like there is a place to put a
socail media images, from settings. Either way, the scss and images
may be useful. The Gemfile is there so that we can build the pages
from localhost. I thought this would cut down on delay and improve
the design process.